Modeled and added platforms
I modeled a platform in blockbench as I tried a scaled cube and it didnt play nice. Ive made it a prefab, but for some reason they are slightly bouncy when interacting with the player. Help with that would be appreciated :)

I also added an extra line to the .gitignore to avoid adding macos things :D